Hauling fixes
